refactor: split safari ai modules
